% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_thresholds_CTfilter}
\alias{calculate_thresholds_CTfilter}
\title{Calculate thresholds for CTfilter}
\usage{
calculate_thresholds_CTfilter(
  ref,
  markers = NULL,
  quant = 0.995,
  assay = "RNA",
  min.sd = 0.02,
  level = 1,
  rm.existing = TRUE,
  method = c("UCell", "AUCell", "ModuleScore"),
  chunk.size = 1000,
  ncores = 1,
  seed = 123,
  verbose = TRUE,
  maxRank = 1500
)
}
\arguments{
\item{ref}{Seurat object containing the reference data set}

\item{markers}{List of markers for each cell type, for example \code{CTfilter::MCA.markers.Mm}}

\item{quant}{Quantile cutoff for score distribution}

\item{assay}{Seurat assay to use}

\item{min.sd}{Minimum value for standard deviation - set to this value if calculated standard deviation is lower}

\item{level}{Annotation level - thresholds are saved in list element \code{ref@misc$CTfilter[[level]]}}

\item{rm.existing}{Overwrite existing CTfilter scores in query object}

\item{method}{Scoring method for cell signatures (default \code{UCell})}

\item{chunk.size}{Number of cells per batch to be scored by the method}

\item{ncores}{Number of processors for parallel processing (requires \code{future.apply})}

\item{seed}{Integer seed for random number generator}

\item{verbose}{Verbose output}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed (used when 'method' is 'UCell' or 'AUCell')}
}
\value{
Return the \code{ref} reference object, with celltype-specific thresholds in the field \code{ref@misc$CTfilter}. Scores for individual signatures are
    returned as metadata in the Seurat object
}
\description{
Given a reference set, calculate thresholds that identify outliers from the reference
}
\examples{
library(ProjecTILs)
ref <- load.reference.map()
ref <- calculate_thresholds_CTfilter(ref)
ref@misc$CTfilter
}
\seealso{
\code{\link{CTfilter()}} to apply signatures on a query dataset and filter on a specific cell type
}
