% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_thresholds_scGate}
\alias{calculate_thresholds_scGate}
\title{Calculate thresholds for scGate}
\usage{
calculate_thresholds_scGate(
  ref,
  markers = NULL,
  quant = 0.995,
  assay = "RNA",
  min.gene.frac = 0.5,
  min.sd = 0.02,
  level = 1,
  rm.existing = TRUE,
  method = c("UCell", "ModuleScore"),
  chunk.size = 1000,
  ncores = 1,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{ref}{Seurat object containing the reference data set}

\item{markers}{List of markers for each cell type, for example \code{scGate::MCA.markers.Mm}}

\item{quant}{Quantile cutoff for score distribution}

\item{assay}{Seurat assay to use}

\item{min.gene.frac}{Only consider signatures covered by this fraction of genes in query set}

\item{min.sd}{Minimum value for standard deviation - if calculated standard deviation is lower than min.sd, it is set to min.sd}

\item{level}{Annotation level - thresholds are saved in list element \code{ref@misc$scGate[[level]]}}

\item{rm.existing}{Overwrite existing scGate scores in query object}

\item{method}{Scoring method for cell signatures (default \code{UCell})}

\item{chunk.size}{Number of cells per batch to be scored by the method}

\item{ncores}{Number of processors for parallel processing (requires \code{future.apply})}

\item{seed}{Integer seed for random number generator}

\item{verbose}{Verbose output}
}
\value{
Return the \code{ref} reference object, with celltype-specific thresholds in the field \code{ref@misc$scGate}. Scores for individual signatures are
    returned as metadata in the Seurat object
}
\description{
Given a reference set, calculate expected mean and SD for all cell types in the reference set. Deviations from this expected distribution (Z-score)
can be used to identify outliers
}
\examples{
library(ProjecTILs)
ref <- load.reference.map()
ref <- calculate_thresholds_scGate(ref)
ref@misc$scGate[[1]]
}
\seealso{
\code{\link{scGate()}} to apply signatures on a query dataset and filter on a specific cell type
}
