% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{combine_scGate_multiclass}
\alias{combine_scGate_multiclass}
\title{Combine scGate annotations}
\usage{
combine_scGate_multiclass(
  obj,
  prefix = "is.pure_",
  scGate_classes = NULL,
  min_cells = 20,
  multi.asNA = FALSE,
  out_column = "scGate_multi"
)
}
\arguments{
\item{obj}{Seurat object with scGate results for multiple models stored as metadata}

\item{prefix}{Prefix in metadata column names for scGate result models}

\item{scGate_classes}{Vector of scGate model names. If NULL, use all columns that start with "prefix" above.}

\item{min_cells}{Minimum number of cells for a cell label to be considered}

\item{multi.asNA}{How to label cells that are "Pure" for multiple annotations: "Multi" (FALSE) or NA (TRUE)}

\item{out_column}{The name of the metadata column where to store the multi-class cell labels}
}
\description{
If a single-cell dataset has precomputed results for multiple scGate models, combined them in multi-class annotation
}
