% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CTfilter.multilevel}
\alias{CTfilter.multilevel}
\title{Filter single-cell data by cell type}
\usage{
CTfilter.multilevel(
  query,
  celltype = "T.cell",
  CT.thresholds = NULL,
  markers = NULL,
  max.impurity = 0.5,
  assay = "RNA",
  min.gene.frac = 0.5,
  sd.in = 2,
  sd.out = 7,
  seed = 1234,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{query}{Seurat object containing a query data set - filtering will be applied to this object}

\item{celltype}{Cell type to preserve from the query data set (should be one of cell types in \code{names(markers)}). For multi-level filtering,
provide nested subtypes as a vector (e.g. \code{celltype=c("T.cell","Tcell.CD8")})}

\item{CT.thresholds}{A named list with thresholds for each cell type - see function \code{\link{calculate_thresholds_CTfilter}}}

\item{markers}{List of markers for each cell type, for example \code{CTfilter::MCA.markers.Mm}}

\item{max.impurity}{Maximum fraction of impurity allowed in clusters to flag cells as "pure"}

\item{assay}{Seurat assay to use}

\item{min.gene.frac}{Only consider signatures covered by this fraction of genes in query set}

\item{sd.in}{Maximum standard deviations from mean (Z-score) to identify outliers for selected signature}

\item{sd.out}{Maximum standard deviations from mean (Z-score) to identify outliers for all other signatures}

\item{seed}{Random seed for cluster analysis}

\item{quiet}{Suppress all output}

\item{...}{Additional parameters for \code{\link{CTfilter}}}
}
\value{
A new metadata column \code{is.pure} is added to the query Seurat object, indicating which cells correspond to the desidered cell type.
    The \code{active.ident} is also set to this variable.
}
\description{
Apply CTfilter for a specific cell type to a query dataset. This function expands \code{\link{CTfilter}} to allow multi-level signatures to be used
for filtering; e.g. Tcells at level 1, and CD8.Tcells at level 2. To prepare markers and signatures in the correct format, store them in a list where
the list index corresponds to the signature level. Then select the desired cell at each level with the celltype parameter
(e.g. \code{celltype=c("T.cell","Tcell.CD8")}). Note that for the parameters listed below, you can also specify different values for different levels, by
setting them as vectors (e.g. sd.out=c(7,4)).
}
\examples{
query <- CTfilter.multilevel(query, celltype=c("T.cell","Tcell.CD4"))
DimPlot(query)
query <- subset(query, subset=is.pure=="Pure")
}
\seealso{
\code{\link{calculate_thresholds_CTfilter()}} to calculate celltype-specific thresholds
}
