% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CTfilter}
\alias{CTfilter}
\title{Filter single-cell data by cell type}
\usage{
CTfilter(
  query,
  celltype = "T.cell",
  CT.thresholds = NULL,
  markers = NULL,
  max.impurity = 0.5,
  ndim = 5,
  resol = 3,
  assay = "RNA",
  genes.blacklist = NULL,
  min.gene.frac = 0.5,
  rm.existing = TRUE,
  seed = 1234,
  skip.normalize = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{query}{Seurat object containing a query data set - filtering will be applied to this object}

\item{celltype}{Cell type to preserve from the query data set (should be one of cell types in \code{names(markers)})}

\item{CT.thresholds}{A named list with thresholds for each cell type - see function \code{\link{calculate_thresholds_CTfilter}}}

\item{markers}{List of markers for each cell type, for example \code{CTfilter::MCA.markers.Mm}}

\item{max.impurity}{Maximum fraction of impurity allowed in clusters to flag cells as "pure"}

\item{ndim}{Number of dimensions for cluster analysis}

\item{assay}{Seurat assay to use}

\item{genes.blacklist}{Genes blacklisted from variable features}

\item{min.gene.frac}{Only consider signatures covered by this fraction of genes in query set}

\item{rm.existing}{Overwrite existing CTfilter scores in query object}

\item{seed}{Random seed for cluster analysis}

\item{skip.normalize}{Skip data normalization}

\item{verbose}{Verbose output}

\item{resul}{Resolution for cluster analysis}
}
\value{
A new metadata column \code{is.pure} is added to the query Seurat object, indicating which cells correspond to the desidered cell type.
    The \code{active.ident} is also set to this variable. Additionally, scores for all signatures in \code{markers} are added to the metadata of the Seurat object.
}
\description{
Apply CTfilter for a specific cell type to a query dataset
}
\examples{
query <- CTfilter(query, celltype="T.cell")
DimPlot(query)
query <- subset(query, subset=is.pure=="Pure")
}
\seealso{
\code{\link{calculate_thresholds_CTfilter()}} to calculate celltype-specific thresholds
}
