% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CTfilter.stats}
\alias{CTfilter.stats}
\title{Calculate thresholds for CTfilter}
\usage{
CTfilter.stats(
  query,
  celltype = "T.cell",
  sd.in = 2,
  sd.out = 7,
  min.cells = 50
)
}
\arguments{
\item{query}{A query object in Seurat format returned by CTfilter}

\item{celltype}{The desidered cell type used in CTfilter}

\item{sd.in}{Maximum standard deviations from mean (Z-score) to identify outliers for selected signature}

\item{sd.out}{Maximum standard deviations from mean (Z-score) to identify outliers for all other signatures}

\item{min.cells}{Minimum number of cells to report a specific cell type. Everything else will be grouped in 'Others'}
}
\value{
Returns the query object with an additional column, reporting the cell type with the highest Z-score compared to the reference
    thresholds. In the slot \code{query@misc$CTfilter.counts}, a dataframe reports the number of cells for each signature that exceed the Z-score threshold. Note that individual
    cells can be outliers for several signatures, so the total count does not correspond to the number of removed cells.
}
\description{
Calculate some stats after a CTfilter run.
}
\examples{
query <- CTfilter(query, celltype="Tcell")
query <- CTfilter.stats(query, celltype="Tcell")
Dimplot(query, group.by="top.zscore")
head(query@misc$CTfilter.counts)
}
\seealso{
\code{\link{CTfilter()}} to apply signatures on a query dataset and filter on a specific cell type
}
