% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{genes.blacklist.default}
\alias{genes.blacklist.default}
\title{Blocklist of genes for dimensionality reduction}
\format{
A list of signatures
}
\description{
A list of signatures, for mouse and human. These include cell cycling,
heat-shock genes, mitochondrial genes, and other genes classes, that may
confound the identification of cell types. These are used internally by scGate
and excluded from the calculation of dimensional reductions (PCA).
}
