% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{modify_model}
\alias{modify_model}
\title{Model signature editing}
\usage{
modify_model(
  model,
  level = NULL,
  name = NULL,
  signature = NULL,
  method = c("replace", "add", "substract")
)
}
\arguments{
\item{model}{scGate model to be modified. When is NULL (default) a new empty model will be initialized.}

\item{level}{integer. It refers to the level of the model tree in wich the signature will be added.}

\item{name}{character indicating signature name (i.e. Immune, TCell, NK etc).}

\item{signature}{character vector indicating genes to be included in the signature. If a minus sign is placed to the end of a gene name, this gene will be used as negative in UCell computing. See UCell documentation for details}

\item{method}{one of  c("replace","add","substract"). When this is set as "replace" (default) the signature will be overwritten. If it set as add, the provided gene(s) will be add to the signature. When you set "substract" the provided genes will be removed of that signature}
}
\description{
Modify a given signature in a scGate existing model
}
