% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.metrics.R
\name{performance.metrics}
\alias{performance.metrics}
\title{Performance metrics}
\usage{
performance.metrics(actual, pred, return_contingency = FALSE)
}
\arguments{
\item{actual}{Logical or numeric binary vector giving the actual cell labels.}

\item{pred}{Logical or numeric binary vector giving predicted cell labels.}

\item{return_contingency}{Return the contingency table? (FALSE)}
}
\value{
Performance metrics (Precision, Recall, MCC) for actual vs predicted
}
\description{
Evaluate model performance for binary tasks
}
\examples{
results <- performance.metrics(actual= sample(c(1,0),20,replace=TRUE),
    pred =  sample(c(1,0),20,replace=TRUE,prob = c(0.65,0.35) ) )

}
