% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Performance metrics}
\usage{
performance_metrics(actual, pred, return_contingency = FALSE)
}
\arguments{
\item{actual}{Logical or numeric binary vector giving the actual cell labels.}

\item{pred}{Logical or numeric binary vector giving the predicted cell labels.}

\item{return_contingency}{Logical indicating if contingency table must be returned. Default is FALSE}
}
\description{
Evaluate model performance for binary tasks
}
\examples{
results <- performance_metrics(actual=sample(c(1,0),20,replace=TRUE),pred=sample(c(1,0),20,replace=TRUE,prob=c(0.65,0.35) ) )
}
