% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_levels}
\alias{plot_levels}
\title{Plot scGate filtering results by level}
\usage{
plot_levels(obj, pure.col = "green", impure.col = "gray")
}
\arguments{
\item{obj}{Gated Seurat object output of scGate filtering function.}
}
\description{
Fast plotting of gating results over each model level.
}
\examples{
library(scGate)
scGate.model.db <- get_scGateDB()
# To see a specific model, browse the list of models:
scGate.model.db$human$generic$Myeloid
# Apply scGate with this model
query <- scGate(query, model=scGate.model.db$human$generic$Myeloid)
plot_levels(query)
}
