% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_tree}
\alias{plot_tree}
\title{View scGate model as a decision tree}
\usage{
plot_tree(model, box.size = 12, edge.text.size = 12)
}
\arguments{
\item{model}{A scGate model to be visualized}

\item{edge.text.size}{Edge text size}

\item{box_size}{Box size}
}
\value{
A plot of the model
}
\description{
View scGate model as a decision tree
}
\examples{
query <- plot_tree(model)
}
