% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{query.seurat}
\alias{query.seurat}
\title{Toy dataset to test the package}
\format{
A Seurat object
}
\description{
A downsampled version (300 cells) of the single-cell dataset by Zilionis et
al. (2019) <doi:10.1016/j.immuni.2019.03.009>, with precalculated PCA and
UMAP reductions.
}
