% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{scGate_Model-class}
\alias{scGate_Model-class}
\alias{scGate_Model}
\title{scGate_Model Class}
\description{
The scGate_Model class is use to store background models generated with the `scGate` package. It contains all the information needed
to apply the model to a query dataset and filter on the cell type(s) of interest stored in the `positive_celltypes` slot.
}
\section{Slots}{

\describe{
\item{\code{species}}{The species of the training data}

\item{\code{positive_celltypes}}{List of celltypes included in the reference set. These must be one or more from the list of markers (`markers` parameter).
Only the `positive_celltypes` will be gated when the model is applied to a new query dataset}

\item{\code{markers}}{List of markers for all celltypes included in the model}

\item{\code{bg_model}}{A dataframe containing mean and SD calculated on training set for all cell types in `markers` list}

\item{\code{scoring_method}}{The algorithm used to calculate signature scores}
}}

\concept{objects}
