% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{test_my_model}
\alias{test_my_model}
\title{Test your model}
\usage{
test_my_model(
  model,
  testing.version = "hsa.latest",
  custom.dataset = NULL,
  target = NULL,
  PLOT = T
)
}
\arguments{
\item{model}{scGate model in data.frame format}

\item{testing.version}{Character indicating the version of testing tatasets to be used. By default "hsa-latest" will be used. It will be ignored if custom.dataset variable is provied in Seurat format. Check available version in "https://figshare.com/XXXX/".}

\item{custom.dataset}{Seurat object to be used as a testing dataset. For testing purposes, metadata seurat object must contain a column named 'cell_type' to be used as a gold standard. Also a set of positive targets mut be provided in the target variable.}
}
\description{
wraper for fast model testing on 3 sampled datasets
}
\examples{
library(scGate)
scGate.model.db <- get_scGateDB()
# Browse the list of models and select one:
model.panBcell <-  scGate.model.db$human$generic$PanBcell
# Case 1: test the model with available testing datasets
panBcell.performance <- test_my_model(model.panBcell, testing.version = 'hsa.latest',target = "PanBcell")
model.Myeloid <-  scGate.model.db$human$generic$Myeloid
myeloid.performance <- test_my_model(model.Myeloid, testing.version = 'hsa.latest',target = "Myeloid")
# Case 2: use your own dataset for testing purposes. 
your.own.seurat.object <- readRDS(path.to.your.custom.dataset)
your.own.seurat.object$cell_type <- your.own.seurat.object$your.gold.standard.column   ## This make a copy of the cell_type reference column as required for scGate
performance.on.custom.dataset <- test_my_model(your.custom.PanBcell.model, custom.dataset = your.own.seurat.object, target = c("Bcell","PlasmaCell"))  # notice that 'Bcell' and 'PlasmaCell' must be celltypes present in your custom.dataset
}
