% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{train_scGate}
\alias{train_scGate}
\title{Generate a scGate model}
\usage{
train_scGate(
  ref,
  markers = NULL,
  positive_celltypes = NULL,
  autocomplete = T,
  maxRank = 1500,
  min.sd = 0.02,
  method = c("UCell", "AUCell", "ModuleScore"),
  quant = 0.995,
  assay = "RNA",
  chunk.size = 1000,
  ncores = 1,
  seed = 123,
  verbose = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{ref}{Seurat object containing the reference data set}

\item{markers}{List of markers for each cell type, for example \code{scGate_DB$human$Tcell@markers}}

\item{positive_celltypes}{List of celltypes included in the reference set. These must be one or more from the list of markers (`\code{markers} parameter).
Only the `positive_celltypes` will be gated when the model is applied to a new query dataset}

\item{autocomplete}{Automatically autocomplete cell types in \code{positive_celltypes} that start with same prefix
(e.g. B.cell gates for B.cell.1, B.cell.2 and B.cell.Plasmocyte signatures)}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed (used when 'method' is 'UCell' or 'AUCell')}

\item{min.sd}{Minimum value for standard deviation - if calculated standard deviation is lower than min.sd, it is set to min.sd}

\item{method}{Scoring method for cell signatures (default \code{UCell})}

\item{quant}{Quantile cutoff for score distribution}

\item{assay}{Seurat assay to use}

\item{chunk.size}{Number of cells per batch to be scored by the method}

\item{ncores}{Number of processors for parallel processing (requires \code{future.apply})}

\item{seed}{Integer seed for random number generator}

\item{verbose}{Verbose output}

\item{quiet}{Suppress all STDOUT output}
}
\value{
Returns a \code{scGate_Model} object, containing the expected signature score distribution (background model) for all evaluated signatures.
}
\description{
Given a reference set, calculate expected mean and SD for all cell types in the reference set. Deviations from this expected distribution (Z-score)
can then be used to gate for specific cell types in any query dataset (see function \code{scGate})
}
\examples{
library(ProjecTILs)
ref <- load.reference.map()
scGate.model <- train_scGate(ref, positive_celltypes='T.cell')
head(scGate.model@bg_model)
}
\seealso{
\code{\link{scGate()}} to apply signatures on a query dataset and filter on specific cell type(s)
}
