% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_thresholds_CTfilter}
\alias{calculate_thresholds_CTfilter}
\title{Calculate thresholds for CTfilter}
\usage{
calculate_thresholds_CTfilter(
  ref,
  markers = NULL,
  sd.dev = 7,
  quant = 0.995,
  assay = "RNA",
  min.gene.frac = 0.5,
  rm.existing = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ref}{Seurat object containing the reference data set}

\item{markers}{List of markers for each cell type, for example `CTfilter::MCA.markers.Mm`}

\item{sd.dev}{Maximum standard deviations from mean to identify outliers}

\item{quant}{Quantile cutoff for score distribution}

\item{assay}{Seurat assay to use}

\item{min.gene.frac}{Only consider signatures covered by this fraction of genes in query set}

\item{rm.existing}{Overwrite existing CTfilter scores in query object}

\item{verbose}{Verbose output}
}
\value{
A new metadata column `is.pure` is added to the query Seurat object, indicating which cells correspond to the desidered cell type.
    The `active.ident` is also set to this variable. Additionally, scores for all signatures in `markers` are added to the metadata of the Seurat object.
}
\description{
Given a reference set, calculate thresholds that identify outliers from the reference
}
\examples{
library(ProjecTILs)
ref <- load.reference.map()
ref <- calculate_thresholds_CTfilter(ref)
ref@misc$CTfilter
}
\seealso{
[CTfilter()] to apply signatures on a query dataset and filter on a specific cell type
}
