% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_thresholds_CTfilter}
\alias{calculate_thresholds_CTfilter}
\title{Calculate thresholds for CTfilter}
\usage{
calculate_thresholds_CTfilter(
  ref,
  markers = NULL,
  quant = 0.995,
  assay = "RNA",
  min.gene.frac = 0.5,
  min.sd = 0.1,
  level = 1,
  rm.existing = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ref}{Seurat object containing the reference data set}

\item{markers}{List of markers for each cell type, for example \code{CTfilter::MCA.markers.Mm}}

\item{quant}{Quantile cutoff for score distribution}

\item{assay}{Seurat assay to use}

\item{min.gene.frac}{Only consider signatures covered by this fraction of genes in query set}

\item{min.sd}{Minimum value for standard deviation - set to this value if calculated standard deviation is lower}

\item{level}{Annotation level - thresholds are saved in list element \code{ref@misc$CTfilter[[level]]}}

\item{rm.existing}{Overwrite existing CTfilter scores in query object}

\item{verbose}{Verbose output}
}
\value{
Return the \code{ref} reference object, with celltype-specific thresholds in the field \code{ref@misc$CTfilter}. Scores for individual signatures are
    returned as metadata in the Seurat object
}
\description{
Given a reference set, calculate thresholds that identify outliers from the reference
}
\examples{
library(ProjecTILs)
ref <- load.reference.map()
ref <- calculate_thresholds_CTfilter(ref)
ref@misc$CTfilter
}
\seealso{
\code{\link{CTfilter()}} to apply signatures on a query dataset and filter on a specific cell type
}
