% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_scGateDB}
\alias{get_scGateDB}
\title{Download scGate model database}
\usage{
get_scGateDB(
  destination = ".",
  force_update = FALSE,
  version = "latest",
  repo_url = "https://github.com/carmonalab/scGate_models"
)
}
\arguments{
\item{destination}{destination path for db storage. The default is current location.}

\item{force_update}{Whether to update an existing database. WARNING: note that setting it TRUE, the current models folder will be overwritten .}

\item{version}{Specify the version of the scGate_models database (e.g. 'v0.1'). By default downloads the latest available version.}

\item{repo_url}{URL path to scGate model repository database}
}
\description{
Download or update the database of scGate models. These are stored in a GitHub repository, from where you can download specific versions of the database.
}
\examples{
scGate.model.db <- get_scGateDB()
}
