% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_scGateDB}
\alias{get_scGateDB}
\title{Load scGate DB models}
\usage{
get_scGateDB(
  destination = ".",
  update_db = F,
  repo_url_zip = "https://github.com/carmonalab/scGate_models/archive/master.zip"
)
}
\arguments{
\item{destination}{destination path for db storage. The default is current location.}

\item{update_db}{Logical indicating if model db must be updated. WARNING: notice that setting it TRUE, the complete folder and models would be overwritten .}

\item{repo_url_zip}{url path to scGate model repository database}
}
\description{
Load scGate DB models
}
\examples{
scGate.model.db <- get_scGateDB()
}
