% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_models}
\alias{load_models}
\title{Load scGate models}
\usage{
load_models(
  repo_path = "scGate_models-master",
  master.table = "master_table.tsv"
)
}
\arguments{
\item{repo_path}{Path to scGate model folder. If this does not exist, the latest version of scGate_model DB is downloaded from the GitHub repo}

\item{master.table}{File name of the master table (in repo_path folder) that contains cell type signatures}
}
\description{
Loads into memory a database of scGate models. These can be either models from the scGate_models DB (https://github.com/carmonalab/scGate_models), or
custom models by the user
}
\examples{
model.list <- load_models()
}
