% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{performance.metrics}
\alias{performance.metrics}
\title{Evaluate model performance for binary tasks}
\usage{
performance.metrics(actual, pred, return_contingency = F)
}
\arguments{
\item{actual}{Logical or numeric binary vector giving the actual cell labels.}

\item{pred}{Logical or numeric binary vector giving the predicted cell labels.}

\item{return_contingency}{Logical indicating if contingency table must be returned. Default is FALSE}
}
\description{
Evaluate model performance for binary tasks
}
\examples{
results <- performance.metrics(actual= sample(c(1,0),20,replace =T),pred =  sample(c(1,0),20,replace = T,prob = c(0.65,0.35) ) )
}
