% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{performance.metrics}
\alias{performance.metrics}
\title{Evaluate model performance for binary tasks}
\usage{
performance.metrics(actual, pred, return_contingency = F)
}
\arguments{
\item{actual}{Logical or numeric binary vector giving the actual cell labels.}

\item{pred}{Logical or numeric binary vector giving the predicted cell labels.}

\item{return_contingency}{Logical indicating if contingency table must be returned. Default is FALSE
If return_contingency =F (default) this function returns a numeric vector containing c(Precision, Recall, Matthews correlation coefficient)
If return_contingency =T, the function returns a list. "summary" element contains the same (PREC, REC, MCC) vector and the "conting" element contains the contingency table.}
}
\description{
Evaluate model performance for binary tasks
}
