% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scGate.multilevel}
\alias{scGate.multilevel}
\title{Filter single-cell data by cell type (multi-level)}
\usage{
scGate.multilevel(
  query,
  celltype = "T.cell",
  bg.model = NULL,
  markers = NULL,
  max.impurity = 0.5,
  assay = "RNA",
  sd.in = 3,
  sd.out = 7,
  seed = 123,
  quiet = FALSE,
  rm.existing = TRUE,
  ...
)
}
\arguments{
\item{query}{Seurat object containing a query data set - filtering will be applied to this object}

\item{celltype}{Cell type to preserve from the query data set (should be one of cell types in \code{names(markers)}). For multi-level filtering,
provide nested subtypes as a vector (e.g. \code{celltype=c("T.cell","Tcell.CD8")})}

\item{bg.model}{The background expected mean and SD for each cell type - see function \code{\link{calculate_thresholds_scGate}}}

\item{markers}{List of markers for each cell type, for example \code{scGate::MCA.markers.Mm}}

\item{max.impurity}{Maximum fraction of impurity allowed in clusters to flag cells as "pure"}

\item{assay}{Seurat assay to use}

\item{sd.in}{Maximum standard deviations from mean (Z-score) to identify outliers for selected signature}

\item{sd.out}{Maximum standard deviations from mean (Z-score) to identify outliers for all other signatures}

\item{seed}{Integer seed for random number generator}

\item{quiet}{Suppress all output}

\item{rm.existing}{Overwrite existing scGate scores in query object}

\item{...}{Additional parameters for \code{\link{scGate}}}
}
\value{
A new metadata column \code{is.pure} is added to the query Seurat object, indicating which cells correspond to the desidered cell type.
    The \code{active.ident} is also set to this variable.
}
\description{
Apply scGate for a specific cell type to a query dataset. This function expands \code{\link{scGate}} to allow multi-level signatures to be used
for filtering; e.g. Tcells at level 1, and CD8.Tcells at level 2. To prepare markers and signatures in the correct format, store them in a list where
the list index corresponds to the signature level. Then select the desired cell at each level with the `celltype` parameter, separating nested subtypes 
using the underscore: e.g. \code{celltype="T.cell_Tcell.CD8"}.
Note that for the parameters listed below, you can also specify different values for different levels, by
setting them as vectors (e.g. sd.out=c(7,4)).
}
\examples{
query <- scGate.multilevel(query, celltype="T.cell_CD4Tcell")
DimPlot(query)
query <- subset(query, subset=is.pure=="Pure")
}
\seealso{
\code{\link{calculate_thresholds_scGate()}} to calculate celltype-specific thresholds
}
