% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scGate}
\alias{scGate}
\title{Filter single-cell data by cell type}
\usage{
scGate(
  data,
  model,
  pos.thr = 0.2,
  neg.thr = 0.2,
  assay = "RNA",
  ncores = 1,
  seed = 123,
  min.cells = 30,
  nfeatures = 2000,
  pca.dim = 30,
  resol = 3,
  output.col.name = "is.pure",
  by.knn = TRUE,
  k.param = 10,
  genes.blacklist = "default",
  additional.signatures = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{Seurat object containing a query data set - filtering will be applied to this object}

\item{pos.thr}{Minimum UCell score value for positive signatures}

\item{neg.thr}{Maximum UCell score value for negative signatures}

\item{assay}{Seurat assay to use}

\item{ncores}{Number of processors for parallel processing (requires \code{future.apply})}

\item{seed}{Integer seed for random number generator}

\item{min.cells}{Stop iterating if fewer than this number of cells is left}

\item{nfeatures}{Number of variable genes for dimensionality reduction}

\item{pca.dim}{Number of principal components for dimensionality reduction}

\item{resol}{Resolution for cluster analysis (if \code{by.knn=FALSE})}

\item{output.col.name}{Column name with 'pure/impure' annotation}

\item{by.knn}{Perform k-nearest neighbor smoothing for UCell scores}

\item{k.param}{Number of nearest neighbors for knn smoothing}

\item{genes.blacklist}{Genes blacklisted from variable features. The default loads the list of genes in \code{scGate::genes.blacklist.default};
you may deactivate blacklisting by setting \code{genes.blacklist=NULL}}

\item{additional.signatures}{A list of additional signatures, not included in the model, to be evaluated (e.g. a cycling signature). The scores for this
list of signatures will be returned but not used for filtering.}

\item{verbose}{Verbose output}

\item{gating.model}{A tabular model with scGate signatures. See description for this format}
}
\value{
A new metadata column \code{is.pure} is added to the query Seurat object, indicating which cells correspond to the desidered cell type(s).
    The \code{active.ident} is also set to this variable.
}
\description{
Apply scGate to filter specific cell types in a query dataset
}
\examples{
query <- scGate(query, model=Bcell.model)
DimPlot(query)
query <- subset(query, subset=is.pure=="Pure")
}
