% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scGate}
\alias{scGate}
\title{Filter single-cell data by cell type}
\usage{
scGate(
  data,
  gating.model = NULL,
  max.impurity = 0.5,
  ndim = 30,
  resol = 3,
  assay = "RNA",
  sd.in = 3,
  sd.out = 7,
  chunk.size = 1000,
  ncores = 1,
  maxRank = 1500,
  max.iterations = 10,
  stop.iterations = 0.01,
  min.cells = 100,
  genes.blacklist = "Tcell.blacklist",
  seed = 123,
  skip.normalize = FALSE,
  return_signature_scores = TRUE,
  verbose = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{data}{Seurat object containing a query data set - filtering will be applied to this object}

\item{gating.model}{The background expected mean and SD for each cell type - see function \code{\link{train_scGate}}}

\item{max.impurity}{Maximum fraction of impurity allowed in clusters to flag cells as "pure". Can be either:
\itemize{
  \item{Single number between 0 and 1 - the same impurity threshold is applied to all iterations}
  \item{Vector of numbers between 0 and 1 - specific impurity thresholds for successive iterations e.g. \code{max.impurity=c(0.7,0.5,0.3)}}
}}

\item{ndim}{Number of dimensions for cluster analysis}

\item{resol}{Resolution for cluster analysis}

\item{assay}{Seurat assay to use}

\item{sd.in}{Maximum standard deviations from mean (Z-score) to identify outliers for selected signatures}

\item{sd.out}{Maximum standard deviations from mean (Z-score) to identify outliers for all other signatures}

\item{chunk.size}{Number of cells per batch to be scored by the method}

\item{ncores}{Number of processors for parallel processing (requires \code{future.apply})}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed (used when 'method' is 'UCell' or 'AUCell')}

\item{max.iterations}{Maximum number of iterations}

\item{stop.iterations}{Stop iterating if fewer than this fraction of cells were removed in the last iteration}

\item{min.cells}{Stop iterating if fewer than this number of cells is left}

\item{genes.blacklist}{Genes blacklisted from variable features. The default loads the list of genes in \code{scGate::genes.blacklist.Mm};
you may deactivate blacklisting by setting \code{genes.blacklist=NULL}}

\item{seed}{Integer seed for random number generator}

\item{skip.normalize}{Skip data normalization}

\item{return_signature_scores}{Add signature scores and Z-scores to object metadata}

\item{verbose}{Verbose output}

\item{quiet}{Suppress all output}
}
\value{
A new metadata column \code{is.pure} is added to the query Seurat object, indicating which cells correspond to the desidered cell type(s).
    The \code{active.ident} is also set to this variable. Additionally, Z-scores for all signatures in \code{markers} are added to the metadata of the Seurat object.
}
\description{
Apply scGate to filter specific cell types in a query dataset
}
\examples{
query <- scGate(query)
DimPlot(query)
query <- subset(query, subset=is.pure=="Pure")
}
\seealso{
\code{\link{train_scGate()}} to generate celltype-specific models
}
